+++ gke/CHANGELOG.md
@@ -5,7 +5,10 @@
 - Disable istio egress gateway by default
 - Able is enable/disable istio control pane
 - Able is enable/disable istio egress gateway
+- Able is enable/disable cert-manager
+- Able is enable/disable external-dns
 - Able is enable/disable kubernetes-dashboard
+- Able is enable/disable stackdriver-adapter
 
 ## 2018-09-16